add: Make snippets have a max number of views
The user is now able to set a maximum number of views/uses on his
snippets. By default, this number is infinite. The user can turn off the
infinite views by unchecking the 'infinite_views checkbox and by
entering a views number in the 'max_views' input field (it has to be a
number greater than 0). Every time the snippet's page is visited, its
views left decrease, and when it gets to 0, the snippet is deleted from
the database.

Related to #5